# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api de gerenciamento de vendas",
        "title": "Gerenciador de vendas",
        "contact": {}
    },
    "host": "localhost:3333",
    "paths": {
        "/perfil": {
            "post": {
                "description": "cria um perfil descrição",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "perfil"
                ],
                "summary": "Cria um perfil",
                "parameters": [
                    {
                        "format": "entity.Perfil",
                        "description": "payload",
                        "name": "perfil",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Perfil"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Perfil"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "description": "Atualiza um perfil descrição",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "perfil"
                ],
                "summary": "Atualiza um perfil",
                "parameters": [
                    {
                        "format": "entity.Perfil",
                        "description": "payload",
                        "name": "perfil",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Perfil"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Perfil"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Perfil": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}